dodanie makroskładników

+++ TDFit/TDFit/Models/Calorie.cs
@@ -18,9 +18,9 @@ namespace TDFit.Models
         public string Activity { get; set; }
         public int ActivityId { get; set; } // aktywnosc fizyczna
 
-        public int Carbohydrate { get; set; }
-        public int Protein { get; set; }
-        public int Fat { get; set; }
+        public double Carbohydrate { get; set; }
+        public double Protein { get; set; }
+        public double Fat { get; set; }
 
         public double Kcal { get; set; }
     }
